+++ packages/components/.storybook/main.js
@@ -92,3 +92,4 @@ module.exports = {
         return config;
       },
 };
+
