+++ packages/components/src/local/coa-status-board/helpers/gen-coa-data.tsx
@@ -1,59 +1,52 @@
 import React from 'react'
-import { MessageCustom, ForceData, ForceRole, TemplateBodysByKey, ChannelData } from '@serge/custom-types'
+import { MessageCustom, ForceData, ForceRole} from '@serge/custom-types'
 import { isMessageReaded, setMessageState } from '@serge/helpers'
-import { RowWithCollapsibleType } from '../../organisms/data-table/types/props'
+import DataTableProps,{ RowWithCollapsibleType } from '../../organisms/data-table/types/props'
 import { ForceColor } from '..'
 import ChannelCoaMessageDetail from '../../molecules/channel-coa-message-detail'
 import getColumns from './get-columns'
 import { Badge } from '../../atoms/badge'
 import { ROW_WITH_COLLAPSIBLE_TYPE } from '../../organisms/data-table'
-import { CollaborativeMessageStates } from '@serge/config'
+import { CollaborativeMessageStates,EMPTY_CELL, SpecialChannelColumns } from '@serge/config'
 import getAssignees from './assignees'
-import getKey from './get-key'
+import { getKeyCOA } from './get-key'
 import { faEnvelope, faEnvelopeOpen } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { Column } from '../../organisms/data-table/types/props'
+import { capitalize } from 'lodash'
 
 /* Import Stylesheet */
 import styles from '../styles.module.scss'
+import { PropsCOA } from '../types/props'
+import getCoaPermissions from './get-coa-perrmisions'
+import { statusColors } from './status-colors'
 
 /** helper to provide legible version of force & role */
 export const formatRole = (role: ForceRole): string => {
   return role.forceName + '-' + role.roleName
 }
 
-const statusColors: { [property: string]: string } = {
-  [CollaborativeMessageStates.Unallocated]: '#B10303',
-  [CollaborativeMessageStates.PendingReview]: '#434343',
-  [CollaborativeMessageStates.Finalized]: '#007219',
-  [CollaborativeMessageStates.Released]: '#007219',
-  [CollaborativeMessageStates.Closed]: '#ff0000',
-  [CollaborativeMessageStates.BeingEdited]: '#ffc107',
-  [CollaborativeMessageStates.Pending]: '#0366d6'
-}
+
+// todo move to in js file
+
+
 
 interface GenData {
-  data: RowWithCollapsibleType[]
+  dataTableProps: DataTableProps
   unreadMessagesCount: number
 }
 
-export const genData = (
-  messages: MessageCustom[],
-  forces: ForceData[],
-  role: ForceRole,
-  currentWargame: string,
-  templates: TemplateBodysByKey,
-  isUmpire: boolean,
-  channel: ChannelData,
-  canCollaborate: boolean,
-  canReleaseMessages: boolean,
-  canUnClaimMessages: boolean,
-  gameDate: string,
-  isCollaborating: boolean,
-  isObserver: boolean,
-  onChange: (msg: MessageCustom) => void,
+export const genCOAData = (
+  props: PropsCOA,
   handleOpenCollapsible: () => void
 ): GenData => {
   let unreadMessagesCount = 0
+
+  const { messages, forces, role, currentWargame, templates, isUmpire, channel, gameDate, onChange } = props
+  const isObserver = isUmpire
+  const { canCollaborate, canReleaseMessages, canUnClaimMessages } = getCoaPermissions(channel, role)
+   // whether this user should see metadata about the message being edited
+  const isCollaborating = canCollaborate || canReleaseMessages || isUmpire
 
   const assignees: ForceRole[] = getAssignees(channel.participants, forces)
 
@@ -78,7 +71,7 @@ export const genData = (
       status === CollaborativeMessageStates.Finalized ||
       status === CollaborativeMessageStates.Released
 
-    const messageStateKey = getKey(message, canCollaborate, canReleaseMessages, canUnClaimMessages, isFinalised, isCollabEditChannel, isObserver)
+    const messageStateKey = getKeyCOA(message, canCollaborate, canReleaseMessages, canUnClaimMessages, isFinalised, isCollabEditChannel, isObserver)
     const isReaded = isMessageReaded(currentWargame, role.forceName, role.roleName, messageStateKey)
 
     if (!isReaded) unreadMessagesCount++
@@ -161,7 +154,116 @@ export const genData = (
     }
   })
 
-  return { data, unreadMessagesCount }
+  const getListOfOwners = (messages: MessageCustom[]): string[] => {
+    const roles = messages.reduce((filters: string[], message) => {
+      const collab = message.details.collaboration
+      if (collab && collab.owner) {
+        return [
+          ...filters,
+          formatRole(collab.owner)
+        ]
+      } else {
+        return filters
+      }
+    }, [])
+    // just the unique sources
+    return [...new Set(roles)]
+  }
+  
+  const getListOfSources = (messages: MessageCustom[]): string[] => {
+    return messages.reduce((filters: any[], message) => {
+      if (!message.details.from) {
+        console.warn(message, 'message have no from.roleName')
+        return filters
+      }
+      return [...new Set([...filters, message.details.from.roleName])]
+    }, [])
+  }
+  
+  const getListOfStatus = (messages: MessageCustom[]): string[] => {
+    return messages.reduce((filters: any[], message) => {
+      if (!message.details.collaboration) {
+        console.warn(message, 'message have no collaboration.status')
+        return filters
+      }
+      return [...new Set([...filters, message.details.collaboration.status])]
+    }, [])
+  }
+  
+  const getListOfExtraColumn = (messages: MessageCustom[], columnName: string): string[] => {
+    const values = messages.reduce((filters: any[], message) => {
+      if (!message.message[columnName]) {
+        return filters
+      }
+      let fields: any
+      switch (columnName) {
+        case 'LOCATION': {
+          const fieldData = []
+          fields = message.message[columnName]
+          for (const key of Object.keys(fields)) {
+            const location = fields[key].map((item: any) => `${key}-${item.Country}`)
+            fieldData.push(...location)
+          }
+          const newItems: string[] = fieldData.length ? fieldData : [EMPTY_CELL]
+          filters.push(...newItems)
+          return filters.sort((a, b) => a === EMPTY_CELL ? -1 : a - b)
+        }
+        default:
+          return filters
+      }
+    }, [])
+    // de-dupe & return
+    return [...new Set(values)]
+  }
+
+  
+  // collate list of message owners
+  const filtersOwners = getListOfOwners(messages)
+
+  // collate list of sources (From) for messages
+  const filtersRoles = getListOfSources(messages)
+
+  // collate list of sources (Status) for messages
+  const filtersStatus = getListOfStatus(messages)
+
+  // collate list of extra column LOCATION for messages
+  const filtersLocations = getListOfExtraColumn(messages, 'LOCATION')
+
+  const columnHeaders: Column[] = [
+    'ID',
+    {
+      filters: filtersRoles,
+      label: 'From'
+    },
+    'Title',
+    {
+      filters: filtersStatus,
+      label: 'Status'
+    },
+    {
+      filters: filtersOwners,
+      label: 'Owner'
+    },
+    'Updated'
+  ]
+
+  if (channel.collabOptions && channel.collabOptions.extraColumns) {
+    const newCols = channel.collabOptions.extraColumns.map((col: SpecialChannelColumns): string | Column => {
+      if (col === SpecialChannelColumns.LOCATION) {
+        return {
+          filters: filtersLocations,
+          label: capitalize(col)
+        }
+      }
+      return capitalize(col)
+    })
+    columnHeaders.push(...newCols)
+  }
+
+  return { dataTableProps: {
+    data,
+    sort: true,
+    columns: columnHeaders
+  }, unreadMessagesCount }
 }
 
-export default genData
