+++ packages/components/src/local/coa-status-board/types/props.d.ts
@@ -6,7 +6,28 @@ import {
   TemplateBodysByKey
 } from '@serge/custom-types'
 
-export default interface Props {
+import { TYPE_COA, TYPE_RFI } from '..'
+
+export interface PropsRFI {
+  type: typeof TYPE_RFI,
+  rfiMessages: MessageCustom[]
+  roles: string[]
+  channels: Array<ChannelData>
+  onChange?: (nextMessage: MessageCustom) => void
+  /** if this player is from the UMPIRE force */
+  isUmpire: boolean
+  /** if this player has "RFI Manager" attribute in their role */
+  isRFIManager: boolean
+  /** role of logged in player
+   */
+  role: ForceRole
+  /** fires on change message status to readed **/
+  onMessageRead?: (unreadedLeft: number) => void
+  currentWargame: string
+}
+
+export interface PropsCOA {
+  type: typeof TYPE_COA,
   currentWargame: string
   messages: MessageCustom[]
   channel: ChannelData
@@ -29,3 +50,7 @@ export default interface Props {
   /** fires on change message status to readed **/
   onMessageRead?: (unreadedLeft: number) => void
 }
+
+type Props = PropsCOA | PropsRFI
+
+export default Props
